requirements.txt
kivy==2.3.0
kivymd==1.2.0
google-genai==0.2.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2